% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model.R
\name{cdcm_stan_file}
\alias{cdcm_stan_file}
\title{Path to the canonical DCM Stan program}
\usage{
cdcm_stan_file()
}
\value{
A character string giving the full path to the
\code{canonical_dcm.stan} file.
}
\description{
Returns the file path to the CDCM
Stan program bundled with the \code{cdcm} package.
}
\details{
This function is primarily intended for use with
\code{cmdstanr::cmdstan_model()} when compiling the model.
}
\examples{
\dontrun{
stan_file <- cdcm_stan_file()
mod <- cmdstanr::cmdstan_model(stan_file)
}

}
