% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_model.R
\name{meta_analysis_stan_file}
\alias{meta_analysis_stan_file}
\title{Path to the group-level meta-analysis Stan program}
\usage{
meta_analysis_stan_file()
}
\value{
A character string giving the full path to the
\code{meta_analysis.stan} file.
}
\description{
Returns the file path to the group-level meta analysis model
Stan program bundled with the \code{cdcm} package.
}
\details{
This function is primarily intended for use with
\code{cmdstanr::cmdstan_model()} when compiling the model.
}
\examples{
\dontrun{
stan_file <- met_analysis_stan_file()
mod <- cmdstanr::cmdstan_model(stan_file)
}

}
