% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_dcm_functions.R
\name{minESS_criterion}
\alias{minESS_criterion}
\title{Compute a minimum ESS threshold for CDCM sampling}
\usage{
minESS_criterion(data, alpha = 0.05, eps = 0.05)
}
\arguments{
\item{data}{A named list of data in the format required by the Stan model,
typically produced by \code{\link{get_stan_dat}}.}

\item{alpha}{Significance level used in the \code{mcmcse::minESS()}
calculation. Defaults to \code{0.05}.}

\item{eps}{Relative precision target used in the
\code{mcmcse::minESS()} calculation. Defaults to \code{0.05}.}
}
\value{
A numeric scalar giving the minimum ESS threshold.
}
\description{
Computes the minimum effective sample size (ESS) threshold corresponding to a
specified confidence level and relative precision, using the number of model
parameters implied by a CDCM Stan data list.
}
\details{
This function first determines the number of model parameters implied by the
supplied Stan data list, then uses \code{mcmcse::minESS()} to compute the
minimum effective sample size required to achieve the specified level
\code{alpha} and relative precision \code{eps}.

This can be used to define the \code{ess_check} argument in
\code{\link{dcm_sample}}.
}
\examples{
\dontrun{
stan_dat <- get_stan_dat(data, hypothesis_idxs)
ess_target <- minESS_criterion(stan_dat)
}

}
